requirements.txt
Flask==3.1.0
redis==5.2.0
google-cloud-storage
